[*] Move get functions into separate file
